refactor : Dish 패키지와 클래스명을 Item으로 변경
