+++ be/src/main/java/sidedish/domain/order/Order.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.relational.core.mapping.Embedded;
-import sidedish.domain.dish.Dish;
 
 @Getter
 @NoArgsConstructor
